feat: purescript
